Create solution.py
This is a brute force method to solve this problem. A more efficient approach can be followed for the prime or not function to reduce the runtime complexity in the best/middle case. The worse case will always be O(N^1/2) for each number and O(N^1/2)*O(index) for the entire program.

The best/middle case runtime can be reduced by: 
1. Checking the 10th place of number in primeornot() and returning False if it equals 0, 2, 4, 6, 8
2. Converting the number to string, then summing them while individually covering them to int. If the sum % 3 ==0 is True, return False (not a prime)
3. If the number is too big, take square root of the number using sqrt() function, compare it with ever maller prime number to see if it is divisible. If not we can confirm that the number is prime.