My solution to the 10001st prime challenge.
The simple solution iterates through every integer ( i ) starting from the first prime (which we know is 2) and checks if i is prime by iterating through every integer ( j ) starting from 2 until the current number ( i ) is reached. For each value of j we check whether there is any remainder left after the division of i and j. If no remainder is left for any value of j we know the current number ( i ) is not prime.  We store each prime we find in a list. We iterate i until the list reaches 10001 elements and return the 10001st element.